refactor modules with scene definition / display to astk to allow re-use in different contexts
